+++ CHANGELOG.md
@@ -13,6 +13,7 @@ This changelog only contains the changes that are unreleased. For changes for in
 - Clarify the error popup after launch when needing Java 16 on Java 8 Minecraft
 - Change text when installing loader to indicate it may take some time
 - Make some dialog windows resize as needed to fit the content
+- Disable Discord integration for Arm devices
 
 ### Misc
 - Change translations using loader/mod platforms to use placeholders